requirements.txt
streamlit
python-dotenv
requests
pypdf
python-dotenv
